+++ include/net/vudpclient.cpp
@@ -33,7 +33,7 @@ bool VUdpClient::doOpen()
   udpSession->handle = socket(AF_INET, SOCK_DGRAM, 0);
   if (udpSession->handle == INVALID_SOCKET)
   {
-    SET_ERROR(VSocketError, "error in socket", WSAGetLastError());
+    SET_ERROR(VSocketError, "error in socket", (int)WSAGetLastError());
     return false;
   }
 
@@ -60,7 +60,7 @@ bool VUdpClient::doOpen()
   int res = bind(udpSession->handle, (struct sockaddr*)&udpSession->addr, sizeof(udpSession->addr));
   if (res == SOCKET_ERROR)
   {
-    SET_ERROR(VSocketError, QString("error in bind(%1:%2)").arg(localHost).arg(localPort), WSAGetLastError());
+    SET_ERROR(VSocketError, QString("error in bind(%1:%2)").arg(localHost).arg(localPort), (int)WSAGetLastError());
     return false;
   }
